bug fixes after trying as a new user

+++ bin/setup.bash
@@ -4,28 +4,28 @@ source ./settings.bash
 source ./etc/functions.bash
 
 #########################################
-#		     Ensure Directories				#
+#                    Ensure Directories                         #
 #########################################
 DelegatorDirectories=(
-	./config_history/
-	./deps/share
-	./deps/src
-	./env/
-	./logs/
-	./mounts/containerUser/
-	./mounts/sysetc/
-	./mounts/sysbin/
-	./input-output/inputs/
-	./input-output/tests/
-	./input-output/outputs/
-	./input-output/work/
-	)
+        ./config_history/
+        ./deps/share
+        ./deps/src
+        ./env/
+        ./logs/
+        ./mounts/containerUser/
+        ./mounts/sysetc/
+        ./mounts/sysbin/
+        ./input-output/inputs/
+        ./input-output/tests/
+        ./input-output/outputs/
+        ./input-output/work/
+        )
 for directory in ${DelegatorDirectories[@]} ; do
-	check_make_directory ${directory}
+        check_make_directory ${directory}
 done
 
 #########################################
-#			Delegator Setup				#
+#                       Delegator Setup                         #
 #########################################
 echo """${FILE_MESSAGE}
 GCC_COLORS='error=01;31:warning=01;35:note=01;36:caret=01;32:locus=01:quote=01'
@@ -40,11 +40,11 @@ echo """#!/usr/bin/env bash
 #########  Added by GLYCAM Delegator Docker setup  #########
 export PATH=/programs/bin:/sysbin:/programs/gems/bin:\${PATH}
 if [[ \"\${PYTHONPATH}\" == *\"/programs/gems\"* ]] ; then
-	:
+        :
 elif [ -z \"\${PYTHONPATH}\" ] ; then
-	PYTHONPATH=/programs/gems
+        PYTHONPATH=/programs/gems
 else
-	PYTHONPATH=/programs/gems:\${PYTHONPATH}
+        PYTHONPATH=/programs/gems:\${PYTHONPATH}
 fi
 export PYTHONPATH
 ############################################################
@@ -62,11 +62,11 @@ export OVERWRITE_OUTPUT=\"False\"
 export UNMIN_SUFFIX=\"_un-min\"
 export DATE_BUILD_DATABASE=\"True\"
 
-# EXIT_SUCESS
-exit 0
 """ > ./mounts/sysetc/_autogen_config.bash
 
-ln -s ./mounts/sysetc/_autogen_config.bash
+if [ ! -e "./_autogen_config.bash" ] ; then
+        ln -s ./mounts/sysetc/_autogen_config.bash
+fi
 
 echo """
 ## SessionSettings.bash
@@ -76,6 +76,11 @@ echo """
 ## Also see: ./docs/Configuration.md
 """ > ./deps/share/SessionSettings.bash
 
-ln -s ./deps/share/SessionSettings.bash
+if [ ! -e "./SessionSettings.bash" ] ; then
+        ln -s ./deps/share/SessionSettings.bash
+fi
 
-ln -s ./mounts/sysetc etc
+if [ ! -e "./etc" ] ; then
+        ln -s ./mounts/sysetc etc
+fi
+
